+++ src/app/locales/en/components/file-browser.js
@@ -8,6 +8,7 @@ import fbRenameModal from './file-browser/fb-rename-modal';
 import fbInfoModal from './file-browser/fb-info-modal';
 import fbDownloadModal from './file-browser/fb-download-modal';
 import fbShareModal from './file-browser/fb-share-modal';
+import fbMetadataModal from './file-browser/fb-metadata-modal';
 
 export default {
   menuCurrentDir: 'Current directory',
@@ -43,4 +44,5 @@ export default {
   fbInfoModal,
   fbDownloadModal,
   fbShareModal,
+  fbMetadataModal,
 };