+++ src/app/routes/application.js
@@ -29,8 +29,6 @@ export default OnedataApplicationRoute.extend(DevelopmentModelRouteMixin, {
   },
 
   beforeModel() {
-    // FIXME:
-    console.log('ver 2');
     const superResult = this._super(...arguments);
     return this.get('onedataWebsocket.webSocketInitializedProxy')
       .catch(() => {