+++ CHANGELOG.md
@@ -6,4 +6,5 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-5767 Added metadata modal
 * VFS-5988 Added shares views and management